+++ tests/host/wasip2-via-wasip3/cli.test.ts
@@ -5,11 +5,11 @@
  * Mirrors wasip2/cli.test.ts.
  */
 
-import { createWasiP2ViaP3Adapter } from './index';
-import { createMockP3 } from './test-helpers';
-import { createStreamPair } from '../wasip3/streams';
-import type { WasiStreamReadable } from '../wasip3/streams';
-import type { WasiInputStream, WasiOutputStream } from './io';
+import { createWasiP2ViaP3Adapter } from '../../../src/host/wasip2-via-wasip3/index';
+import { createMockP3 } from '../../../src/host/wasip2-via-wasip3/test-helpers';
+import { createStreamPair } from '../../../src/host/wasip3/streams';
+import type { WasiStreamReadable } from '../../../src/host/wasip3/streams';
+import type { WasiInputStream, WasiOutputStream } from '../../../src/host/wasip2-via-wasip3/io';
 
 describe('wasi:cli/environment (via P3 adapter)', () => {
     it('getEnvironment returns configured pairs from mock', () => {
