+++ tests/marshal/resources.test.ts
@@ -1,16 +1,16 @@
 // Copyright (c) 2023 Pavel Savara. Licensed under the MIT License.
 
-import { initializeAsserts } from '../utils/assert';
+import { initializeAsserts } from '../../src/utils/assert';
 initializeAsserts();
 
-import { ModelTag } from '../parser/model/tags';
-import { PrimitiveValType } from '../parser/model/types';
-import { ResolverContext, MarshalingContext } from '../resolver/types';
-import { createResourceTable } from '../runtime';
-import { resolveCanonicalResourceType } from '../resolver/type-resolution';
-import { createLifting as _createLifting } from '../binder/to-abi';
-import { createLowering } from '../binder/to-js';
-import type { WasmValue } from './model/types';
+import { ModelTag } from '../../src/parser/model/tags';
+import { PrimitiveValType } from '../../src/parser/model/types';
+import { ResolverContext, MarshalingContext } from '../../src/resolver/types';
+import { createResourceTable } from '../../src/runtime';
+import { resolveCanonicalResourceType } from '../../src/resolver/type-resolution';
+import { createLifting as _createLifting } from '../../src/binder/to-abi';
+import { createLowering } from '../../src/binder/to-js';
+import type { WasmValue } from '../../src/marshal/model/types';
 import { describeDebugOnly } from '../test-utils/debug-only';
 
 // Wrap BYO-buffer lifters to return arrays for test convenience
