+++ tests/parser/export.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2023 Pavel Savara. Licensed under the MIT License.
 
-import { expectModelToEqualWat } from './jest-utils';
-import { ComponentExternalKind } from './model/exports';
-import { ModelTag } from './model/tags';
+import { expectModelToEqualWat } from '../../src/parser/jest-utils';
+import { ComponentExternalKind } from '../../src/parser/model/exports';
+import { ModelTag } from '../../src/parser/model/tags';
 
 describe('export', () => {
     test('parse export', async () => {
